Add new show route
Add new actions for deleting, and getting specific posts.

Changed promises chaining with await

Add callbacks to delete and posts new components so that whenever an app creates or deletes a posts it will go to the index page as long as the request is successful

Made usage of already loaded posts to not have to fetch the post if it was already previously loaded by the index route.

Removed the post from the state whenever a user clicks on delete post, so that the app renders a faster new posts index and then it will refresh the state by calling the index posts endpoint